Part of Level class, background showable
Added a background folder in content where background pngs will be kept.
Level class is used to load images of these pngs onto the background of
the game window. MyChar.png was modified using GIMP so that the
background of the character sprites are now transparent (At first, Quote
was Quote + a black square on the background, now it's just Quote).

[ticket: X]