Add tinyxml lib, add first map file
Used Tiled Map Editor from Itch.io to put together the first map of the
Cavestory game using the .png tileset from the cavestory data. Added
this map to the maps folder in content. Added tinyxml2 source and header
to respective folders for use in further episodes of the series.

[ticket: X]