tile class
Tile class is now used insteaed of just loading a rectangle from
spritesheet over and over again we look at a .tmx file and parse it
using the tinyxml2 library. The .tmx file is a map file which can be
loaded in and parsed for information on the tiles loaded for each 16x16
rectangle of the window. This means that we can now build .tmx map files using a
map editor (Tiled) and use those files in loading up different levels
for the game. Amazing

[ticket: X]