Add rectangle, does not compile
This is me messing around with github to check differences I made in the
files to find out why it's not compiling at all now or why it won't load
the map now. Added collision checking and collision objects to the
map_1.tmx so wondering where it all went wrong.

[ticket: X]