Testing
Trying to get all of it to work. It "randomly" stopped working for some
reason, and backing up the head to the github master doesn't fix it. I
think it has to do with what pathway the compiler chooses for finding
libpng but am not sure. Taking a break for now to come back to it later.

[ticket: X]